refactor: move posting/commenting/replying reads to user/api
