+++ apps/web/src/shared/api/supabaseReads.ts
@@ -16,241 +16,8 @@ export function formatInFilter(values: string[]): string {
   return `(${quoted.join(',')})`;
 }
 
-// Supabase row types for query results
-interface PostRow {
-  id: string;
-  board_id: string;
-  title: string;
-  content_length: number;
-  created_at: string;
-}
-
 // Note: Supabase join selects return arrays at the type level,
 // but !inner guarantees exactly one row. We use type assertions in the mappers.
-
-// Types matching the Firestore fan-out models for compatibility
-export interface SupabasePosting {
-  board: { id: string };
-  post: { id: string; title: string; contentLength: number };
-  createdAt: Date;
-  isRecovered?: boolean;
-}
-
-export interface SupabaseCommenting {
-  board: { id: string };
-  post: { id: string; title: string; authorId: string };
-  comment: { id: string; content: string };
-  createdAt: Date;
-}
-
-export interface SupabaseReplying {
-  board: { id: string };
-  post: { id: string; title: string; authorId: string };
-  comment: { id: string; authorId: string };
-  reply: { id: string };
-  createdAt: Date;
-}
-
-/**
- * Fetch user's posts from Supabase posts table.
- * Replaces: users/{userId}/postings subcollection
- * Uses index: idx_posts_author_created
- */
-export async function fetchPostingsFromSupabase(userId: string): Promise<SupabasePosting[]> {
-  const supabase = getSupabaseClient();
-
-  const { data, error } = await supabase
-    .from('posts')
-    .select('id, board_id, title, content_length, created_at')
-    .eq('author_id', userId)
-    .order('created_at', { ascending: false });
-
-  if (error) {
-    console.error('Supabase fetchPostings error:', error);
-    throw error;
-  }
-
-  return (data || []).map((row: PostRow) => ({
-    board: { id: row.board_id },
-    post: {
-      id: row.id,
-      title: row.title,
-      contentLength: row.content_length ?? 0,
-    },
-    createdAt: new Date(row.created_at),
-  }));
-}
-
-/**
- * Fetch user's posts within a date range from Supabase.
- * Replaces: users/{userId}/postings with date filter
- */
-export async function fetchPostingsByDateRangeFromSupabase(
-  userId: string,
-  start: Date,
-  end: Date
-): Promise<SupabasePosting[]> {
-  const supabase = getSupabaseClient();
-
-  const { data, error } = await supabase
-    .from('posts')
-    .select('id, board_id, title, content_length, created_at')
-    .eq('author_id', userId)
-    .gte('created_at', start.toISOString())
-    .lte('created_at', end.toISOString())
-    .order('created_at', { ascending: false });
-
-  if (error) {
-    console.error('Supabase fetchPostingsByDateRange error:', error);
-    throw error;
-  }
-
-  return (data || []).map((row: PostRow) => ({
-    board: { id: row.board_id },
-    post: {
-      id: row.id,
-      title: row.title,
-      contentLength: row.content_length ?? 0,
-    },
-    createdAt: new Date(row.created_at),
-  }));
-}
-
-/** Shared join fields for posts table */
-interface PostJoinFields {
-  id: string;
-  title: string;
-  author_id: string;
-  board_id: string;
-}
-
-/** Row from: comments + posts!inner join */
-interface CommentWithPostJoin {
-  id: string;
-  content: string;
-  created_at: string;
-  post_id: string;
-  posts: PostJoinFields | PostJoinFields[];
-}
-
-/** Row from: replies + comments!inner + posts!inner join */
-interface ReplyWithJoins {
-  id: string;
-  created_at: string;
-  comment_id: string;
-  post_id: string;
-  user_id: string;
-  comments: { id: string } | { id: string }[];
-  posts: PostJoinFields | PostJoinFields[];
-}
-
-/**
- * Fetch user's comments within a date range from Supabase.
- * Replaces: users/{userId}/commentings subcollection
- * Uses index: idx_comments_user_created
- *
- * Note: Requires JOIN with posts table to get post title and author_id
- */
-export async function fetchCommentingsByDateRangeFromSupabase(
-  userId: string,
-  start: Date,
-  end: Date
-): Promise<SupabaseCommenting[]> {
-  const supabase = getSupabaseClient();
-
-  const { data, error } = await supabase
-    .from('comments')
-    .select(`
-      id,
-      content,
-      created_at,
-      post_id,
-      posts!inner (
-        id,
-        title,
-        author_id,
-        board_id
-      )
-    `)
-    .eq('user_id', userId)
-    .gte('created_at', start.toISOString())
-    .lt('created_at', end.toISOString())
-    .order('created_at', { ascending: false });
-
-  if (error) {
-    console.error('Supabase fetchCommentings error:', error);
-    throw error;
-  }
-
-  return ((data || []) as CommentWithPostJoin[]).map((row) => {
-    const post = Array.isArray(row.posts) ? row.posts[0] : row.posts;
-    return {
-      board: { id: post.board_id },
-      post: { id: post.id, title: post.title, authorId: post.author_id },
-      comment: { id: row.id, content: row.content },
-      createdAt: new Date(row.created_at),
-    };
-  });
-}
-
-/**
- * Fetch user's replies within a date range from Supabase.
- * Replaces: users/{userId}/replyings subcollection
- * Uses index: idx_replies_user_created
- *
- * Note: Uses denormalized post_id on replies table for efficiency
- */
-export async function fetchReplyingsByDateRangeFromSupabase(
-  userId: string,
-  start: Date,
-  end: Date
-): Promise<SupabaseReplying[]> {
-  const supabase = getSupabaseClient();
-
-  // Use compound filter to avoid duplicate created_at params issue
-  const startIso = start.toISOString();
-  const endIso = end.toISOString();
-
-  const { data, error } = await supabase
-    .from('replies')
-    .select(`
-      id,
-      created_at,
-      comment_id,
-      post_id,
-      user_id,
-      comments!inner (
-        id
-      ),
-      posts!inner (
-        id,
-        title,
-        author_id,
-        board_id
-      )
-    `)
-    .eq('user_id', userId)
-    .gte('created_at', startIso)
-    .lt('created_at', endIso)
-    .order('created_at', { ascending: false });
-
-  if (error) {
-    console.error('Supabase fetchReplyings error:', error);
-    throw error;
-  }
-
-  return ((data || []) as ReplyWithJoins[]).map((row) => {
-    const post = Array.isArray(row.posts) ? row.posts[0] : row.posts;
-    const comment = Array.isArray(row.comments) ? row.comments[0] : row.comments;
-    return {
-      board: { id: post.board_id },
-      post: { id: post.id, title: post.title, authorId: post.author_id },
-      comment: { id: comment.id, authorId: '' }, // comment author not available (column ambiguity)
-      reply: { id: row.id },
-      createdAt: new Date(row.created_at),
-    };
-  });
-}
 
 // ================================================
 // Core Entity Read Functions (Phase A migration)
